add debug용 tp채우기 기능 추가

+++ Cronos_URP/Assets/Script/TestEnemyAI_Script/DamageSystem/MeleeTriggerEnterDamager.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [RequireComponent(typeof(Collider))]
 public class MeleeTriggerEnterDamager : MonoBehaviour
 {
-    public int damageAmount = 1;
+    public float damageAmount = 1;
     public bool stopCamera = false;
 
     public LayerMask targetLayers;
